styling and refactoring
