+++ frontend/src/index.css
@@ -22,7 +22,7 @@ code {
   text-align: center;
   color: #2d3748;
   width: 75%;
-  padding-top: 10%;
+  padding-top: 2%;
   margin: 0px auto;
 }
